+++ src/DeVinci_frontend/components/SelectModel.svelte
@@ -1,55 +1,171 @@
-<script lang="ts">
-  import {
-    deviceType
-  } from "../store";
+<script>
+  import logo from '../assets/devinci.svg';
+  import hero from '../assets/apple-touch-icon.png'; // Example image for LLM hero section
+  // import '../index.scss';
+    // import { error } from 'console';
+    //import {goto} from ''
+ // import { authenticateWithInternetIdentity } from '../routes/services/auth';
 
-  import SelectModelOption from './SelectModelOption.svelte';
+  let isMenuOpen = false;
 
-  import { getAvailableAiModels } from "../helpers/ai_model_helpers";
-  import { userHasDownloadedModel } from "../helpers/localStorage";
+  function toggleMenu() {
+    isMenuOpen = !isMenuOpen;
+  }
 
-  export let onlyShowDownloadedModels = false;
-  export let autoInitiateSelectedModel = false;
+  function handleSignUpClick() {
+    goto('/login');
+  }
 
-  // Reactive statement to check if the user has already downloaded at least one AI model
-  $: userHasDownloadedAtLeastOneModel = userHasDownloadedModel();
+  function handleLoginClick() {
+   // authenticateWithInternetIdentity().catch((error) => {
+      // console.error("Authentication failed:", error);
+   // });
+  }
 
-  let availableAiModels = getAvailableAiModels(deviceType === 'Android');
-  let chatModelDownloadInProgress = false;
-
+  const items = [
+    {
+      id: 'item1',
+      title: 'Enhanced Natural Language Processing',
+      description: "Experience the power of advanced LLM that understands and generates human-like text. Elevate your conversational AI and content creation.",
+      imgSrc: logo
+    },
+    {
+      id: 'item2',
+      title: 'Industry-specific Models',
+      description: "Our LLMs are tailored for various industries like healthcare, finance, and legal, providing specialized insights and automation.",
+      imgSrc: logo
+    },
+    {
+      id: 'item3',
+      title: 'Seamless Integration',
+      description: "Integrate LLM technology effortlessly with your existing platforms to enhance user experience, productivity, and creativity.",
+      imgSrc: logo
+    }
+  ];
 </script>
 
-  <div>
-    <div class="m-4 mt-2 px-4 py-3 text-gray-500 inline-block border border-[#b0c4de] rounded-lg sm:px-5 bg-gray-50">
-      <b>Please select a model</b>. When selecting an AI model to interact with, consider the balance between performance and hardware requirements. Larger models generally offer enhanced accuracy and advanced features, making them ideal for complex tasks. However, they demand more powerful hardware and may operate slower.
-    </div>
-    <ul class="grid w-full gap-2 md:grid-cols-2 px-4 mt-4">
-      {#each availableAiModels as option}
-        {#if userHasDownloadedAtLeastOneModel}
-          <SelectModelOption
-            id={option.id}
-            name="{option.name}"
-            value={option.value}
-            title={option.title}
-            parameters={option.parameters}
-            performance={option.performance}
-            size={option.size}
-            bind:chatModelDownloadInProgress
-            onlyShowIfDownloaded={onlyShowDownloadedModels}
-            autoInitiateIfModelSelected={autoInitiateSelectedModel}
-          />
-        {:else}
-          <SelectModelOption
-            id={option.id}
-            name="{option.name}"
-            value={option.value}
-            title={option.title}
-            parameters={option.parameters}
-            performance={option.performance}
-            size={option.size}
-            bind:chatModelDownloadInProgress
-          />
-        {/if}
-      {/each}
-    </ul>
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
+
+<!-- Navbar -->
+<header class="bg-gradient-to-r from-[#00b4d8] to-[#0f4c75] text-white py-4 px-6 flex flex-col md:flex-row items-center justify-between shadow-lg fixed top-0 w-full z-50">
+  <div class="flex items-center space-x-4 w-full md:w-auto">
+    <a href="/" class="flex items-center space-x-2">
+      <img src={logo} alt="LLM Logo" class="w-12 h-12 sm:w-14 sm:h-14 rounded-full border-4 border-blue-200 transition-transform transform hover:scale-110 duration-300" />
+      <h1 class="text-xl sm:text-2xl md:text-3xl font-bold">LLMVerse</h1>
+    </a>
   </div>
+
+  <!-- Center Links (Desktop) -->
+  <nav class="hidden md:flex flex-grow justify-center space-x-6 mt-4 md:mt-0">
+    {#each ['Overview', 'Features', 'Integrations', 'Contact'] as link}
+      <a href="#{link.toLowerCase()}" class="text-white hover:text-blue-300 border-b-2 border-transparent hover:border-white px-4 py-2 transition-all duration-300">{link}</a>
+    {/each}
+  </nav>
+
+  <!-- Right Section Buttons -->
+  <div class="flex space-x-4 mt-4 md:mt-0">
+    <button class="bg-[#023e8a] text-white px-4 py-2 rounded-full shadow-md hover:bg-[#0077b6] transition-all duration-300 transform hover:scale-105">Get Started 🚀</button>
+    <button class="bg-transparent border border-blue-200 text-white px-4 py-2 rounded-full shadow-md hover:bg-blue-300 hover:text-[#023e8a] transition-all duration-300 transform hover:scale-105">Partner with Us 🤝</button>
+  </div>
+</header>
+
+<!-- Hero Section -->
+<section class="hero min-h-screen flex items-center justify-center bg-[#0077b6] text-white">
+  <div class="container mx-auto flex flex-col md:flex-row items-center justify-between p-8">
+    <!-- Left Side: Text Content -->
+    <div class="text-content md:w-1/2 text-center md:text-left">
+      <h1 class="text-4xl md:text-5xl font-bold mb-4 leading-tight">
+        Empower Your Business with <span class="text-[#00C4CC]">LLMVerse</span>
+      </h1>
+      <p class="text-lg md:text-xl mb-6">
+        Unlock the potential of large language models for enhanced customer engagement, automation, and data insights.
+      </p>
+      <!-- Styled Buttons -->
+      <div class="flex justify-center md:justify-start space-x-4">
+        <button 
+          on:click={handleSignUpClick}
+          class="bg-[#00C4CC] text-white px-11 py-2 text-lg font-semibold shadow-lg hover:bg-white hover:text-[#0077b6] transition-all duration-300 ease-in-out transform hover:scale-105">
+          Sign Up
+        </button>
+        <button 
+          on:click={handleLoginClick}
+          class="bg-transparent border-2 border-white text-white px-11 py-2 text-lg font-semibold shadow-lg hover:bg-white hover:text-[#0077b6] transition-all duration-300 ease-in-out transform hover:scale-105">
+          Login
+        </button>
+      </div>
+    </div>
+
+    <!-- Right Side: Image or Graphic -->
+    <div class="image-content md:w-1/2 mt-8 md:mt-0">
+      <img src={hero} alt="LLM Hero Graphic" class="w-full h-auto rounded-lg" />
+    </div>
+  </div>
+</section>
+
+<!-- Features Section -->
+<section id="features" class="section-padding bg-gradient-to-r from-[#f0f4f8] to-white py-16">
+  <div class="container mx-auto text-center">
+    <h2 class="text-4xl font-bold text-[#023e8a] mb-8">Platform Features</h2>
+    <p class="text-gray-600 mb-12 max-w-3xl mx-auto">Discover the cutting-edge capabilities of LLMVerse. Explore industry-specific models, seamless integration, and unmatched language understanding.</p>
+    <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+      {#each items as {title, description} }
+        <div class="bg-white p-6 rounded-lg shadow-lg hover:shadow-xl transition-shadow duration-300 transform hover:scale-105">
+          <h3 class="text-xl font-semibold text-[#023e8a] mb-4">{title}</h3>
+          <p class="text-gray-600">{description}</p>
+        </div>
+      {/each}
+    </div>
+  </div>
+</section>
+
+<!-- Integrations Section -->
+<section id="integrations" class="section-padding bg-gradient-to-r from-blue-50 to-white py-16">
+  <div class="container mx-auto text-center">
+    <h2 class="text-4xl font-bold text-[#023e8a] mb-8">Seamless Integrations</h2>
+    <p class="text-gray-600 mb-12 max-w-2xl mx-auto">
+      Connect with various platforms and enhance the AI-driven features of your products. LLMVerse fits into your existing workflow effortlessly.
+    </p>
+
+    <!-- Integrations Cards -->
+    <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+      {#each items as {imgSrc, title, description} }
+        <div class="bg-white p-6 rounded-lg shadow-lg hover:shadow-xl transition-shadow duration-300 transform hover:scale-105">
+          <img src={imgSrc} alt="Integration" class="w-full h-48 object-cover rounded-md mb-4" />
+          <h3 class="text-xl font-semibold text-[#023e8a] mb-2">{title}</h3>
+          <p class="text-gray-600 mb-4">{description}</p>
+        </div>
+      {/each}
+    </div>
+  </div>
+</section>
+
+<!-- Licensing Section -->
+<section id="licensing" class="section-padding bg-gradient-to-r from-[#f9f9f9] to-white py-16">
+  <div class="container mx-auto text-center">
+    <h2 class="text-4xl font-bold text-[#023e8a] mb-8">Licensing Opportunities</h2>
+    <p class="text-gray-600 mb-12 max-w-3xl mx-auto">
+      Maximize your intellectual property value with licensing solutions from LLMVerse. Flexible, scalable models for every industry.
+    </p>
+
+    <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+      <!-- Licensing cards will go here -->
+    </div>
+  </div>
+</section>
+
+<!-- Footer -->
+<footer class="bg-[#023e8a] text-white py-6 text-center">
+  <div class="container mx-auto">
+    <p class="mb-4">&copy; 2024 LLMVerse. All rights reserved.</p>
+    <div class="space-x-4 mb-4">
+      <a href="/privacy" class="hover:text-blue-300">Privacy Policy</a>
+      <a href="/terms" class="hover:text-blue-300">Terms of Service</a>
+      <a href="#contact" class="hover:text-blue-300">Contact</a>
+    </div>
+    <div class="space-x-6">
+      <a href="#" class="hover:text-blue-300"><i class="fab fa-facebook"></i></a>
+      <a href="#" class="hover:text-blue-300"><i class="fab fa-twitter"></i></a>
+      <a href="#" class="hover:text-blue-300"><i class="fab fa-linkedin"></i></a>
+    </div>
+  </div>
+</footer>
